+++ messages.go
@@ -27,6 +27,7 @@ type MessageCreateParams struct {
 	MessageExpiry       int          `json:"message_expiry,omitempty" url:"message_expiry,omitempty"`
 	Template            *Template    `json:"template,omitempty" url:"template,omitempty"`
 	Interactive         *Interactive `json:"interactive,omitempty" url:"interactive,omitempty"`
+	Location            *Location    `json:"location,omitempty" url:"location,omitempty"`
 	DLTEntityID         string       `json:"dlt_entity_id,omitempty" url:"dlt_entity_id,omitempty"`
 	DLTTemplateID       string       `json:"dlt_template_id,omitempty" url:"dlt_template_id,omitempty"`
 	DLTTemplateCategory string       `json:"dlt_template_category,omitempty" url:"dlt_template_category,omitempty"`
@@ -140,6 +141,14 @@ type Parameter struct {
 	Payload  string    `mapstructure:"payload" json:"payload,omitempty"`
 	Currency *Currency `mapstructure:"currency" json:"currency,omitempty"`
 	DateTime *DateTime `mapstructure:"date_time" json:"date_time,omitempty"`
+	Location Location  `mapstructure:"location" json:"location,omitempty"`
+}
+
+type Location struct {
+	Longitude string `mapstructure:"longitude" json:"longitude,omitempty" validate:"required"`
+	Latitude  string `mapstructure:"latitude" json:"latitude,omitempty" validate:"required"`
+	Name      string `mapstructure:"name" json:"name,omitempty" validate:"required"`
+	Address   string `mapstructure:"address" json:"address,omitempty" validate:"required"`
 }
 
 type Currency struct {